Update Makefile and code files
Update Makefile to change the arguments passed to the program in the 'hell' target.

Update 'death.c' to remove unnecessary code related to unlocking a mutex.

Update 'philo.h' to change the author and update the date.

Update 'print.c' to fix a bug related to accessing a variable.

Update 'threads.c' to fix a bug related to unlocking a mutex and destroying it.